requirements.txt
streamlit
pandas
cx_Oracle
ultralytics
Pillow
boto3